fect 添加npm源的选项
